# backend_in_node/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Products API documentation",
        "version": "1.0.0",
        "title": "Products API",
        "contact": {
            "email": "hectorandres.pv@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:3000",
    "basePath": "/api",
    "paths" : {
...
        "/update_product/{product_id}" : {
            "patch" : {
                "summary" : "Update the product with this id",
                "description": "Update the product with this id",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "product id that needs to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "product object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "product" : {
                                    "type": "object",
                                    "$ref": "#/definitions/productToAddOrEdit"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dataWrittenToJSONFile"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
    }, 
    "definitions": {
        "dataWrittenToJSONFile": {
            "type": "object",
            "properties": {
                "all_products":{
                    "type":"[] Product"
                }
            }
        },
...
        "productToAddOrEdit": {
            "type": "object",
            "properties": {
                "productName": {
                    "type": "string"
                },
                "productOwnerName": {
                    "type": "string"
                },
                "Developers": {
                    "type": "array"
                },
                "scrumMasterName": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "methodology": {
                    "type": "string"
                }  
            }
        },
...
    }
}